Ports: Do not download sources if they're already present
When running ./package.sh to rebuild an already installed port, we would not
want to spend time re-downlodaing the same tarball again. Ideally, this should
use some sort of hash checking to ensure the file is not truncated or something,
but this is good enough for now.

+++ Ports/.port_include.sh
@@ -42,7 +42,11 @@ func_defined fetch || fetch() {
     for f in $files; do
         IFS=$OLDIFS
         read url filename <<< $(echo "$f")
-        run_nocd curl ${curlopts:-} "$url" -o "$filename"
+        if [ -f "$filename" ]; then
+            echo "$filename already exists"
+        else
+            run_nocd curl ${curlopts:-} "$url" -o "$filename"
+        fi
         case "$filename" in
             *.tar*|.tbz*|*.txz|*.tgz)
                 run_nocd tar xf "$filename"
@@ -57,7 +61,7 @@ func_defined fetch || fetch() {
                 echo "Note: no case for file $filename."
                 ;;
         esac
-    done    
+    done
     if [ -d patches ]; then
         for f in patches/*; do
             run patch -p"$patchlevel" < "$f"
